+++ hc/front/tests/test_details.py
@@ -29,7 +29,7 @@ class DetailsTestCase(BaseTestCase):
     def test_it_checks_ownership(self):
         self.client.login(username="charlie@example.org", password="password")
         r = self.client.get(self.url)
-        assert r.status_code == 403
+        self.assertEqual(r.status_code, 404)
 
     def test_it_shows_cron_expression(self):
         self.check.kind = "cron"
@@ -38,3 +38,11 @@ class DetailsTestCase(BaseTestCase):
         self.client.login(username="alice@example.org", password="password")
         r = self.client.get(self.url)
         self.assertContains(r, "Cron Expression", status_code=200)
+
+    def test_it_allows_cross_team_access(self):
+        self.bobs_profile.current_team = None
+        self.bobs_profile.save()
+
+        self.client.login(username="bob@example.org", password="password")
+        r = self.client.get(self.url)
+        self.assertEqual(r.status_code, 200)