+++ hc/front/tests/test_remove_check.py
@@ -9,37 +9,32 @@ class RemoveCheckTestCase(BaseTestCase):
         self.check = Check(user=self.alice)
         self.check.save()
 
-    def test_it_works(self):
-        url = "/checks/%s/remove/" % self.check.code
+        self.remove_url = "/checks/%s/remove/" % self.check.code
 
+    def test_it_works(self):
         self.client.login(username="alice@example.org", password="password")
-        r = self.client.post(url)
+        r = self.client.post(self.remove_url)
         self.assertRedirects(r, "/checks/")
 
-        assert Check.objects.count() == 0
+        self.assertEqual(Check.objects.count(), 0)
 
     def test_team_access_works(self):
-        url = "/checks/%s/remove/" % self.check.code
-
         # Logging in as bob, not alice. Bob has team access so this
         # should work.
         self.client.login(username="bob@example.org", password="password")
-        self.client.post(url)
-        assert Check.objects.count() == 0
+        self.client.post(self.remove_url)
+
+        self.assertEqual(Check.objects.count(), 0)
 
     def test_it_handles_bad_uuid(self):
-        url = "/checks/not-uuid/remove/"
-
         self.client.login(username="alice@example.org", password="password")
-        r = self.client.post(url)
+        r = self.client.post("/checks/not-uuid/remove/")
         self.assertEqual(r.status_code, 404)
 
     def test_it_checks_owner(self):
-        url = "/checks/%s/remove/" % self.check.code
-
         self.client.login(username="charlie@example.org", password="password")
-        r = self.client.post(url)
-        assert r.status_code == 403
+        r = self.client.post(self.remove_url)
+        self.assertEqual(r.status_code, 404)
 
     def test_it_handles_missing_uuid(self):
         # Valid UUID but there is no check for it:
@@ -47,10 +42,17 @@ class RemoveCheckTestCase(BaseTestCase):
 
         self.client.login(username="alice@example.org", password="password")
         r = self.client.post(url)
-        assert r.status_code == 404
+        self.assertEqual(r.status_code, 404)
 
     def test_it_rejects_get(self):
-        url = "/checks/%s/remove/" % self.check.code
         self.client.login(username="alice@example.org", password="password")
-        r = self.client.get(url)
+        r = self.client.get(self.remove_url)
         self.assertEqual(r.status_code, 405)
+
+    def test_it_allows_cross_team_access(self):
+        self.bobs_profile.current_team = None
+        self.bobs_profile.save()
+
+        self.client.login(username="bob@example.org", password="password")
+        r = self.client.post(self.remove_url)
+        self.assertRedirects(r, "/checks/")